Add Weka integration (requires docker-compose build)
Now we can use Weka's classes and methods.

The additional dokerfile/requirements include java, weka, and a few python dependencies.

The java version is an LTS one, but was pulled from the default debian package repo for a security vulnerability - the new lines ask for this version explicitly.

Soon: fixing the classifier input in the final cell (to fix the classifier output).