Add experiment structure for complex bayes
It now records multiple results rather than just the one on screen.
Thoroughly inspired by the q 4/5/6 code for coherence.

Also, up the JVM memory limit. I'm not sure if I was running into this, but its good to know that it shouldn't be the killer issue anymore when something does go wrong.